+++ tests/test_all_the_things/test_all_the_things_suite_test.go
@@ -17,16 +17,19 @@ import (
 	"github.com/k8ssandra/cass-operator/tests/kustomize"
 	ginkgo_util "github.com/k8ssandra/cass-operator/tests/util/ginkgo"
 	"github.com/k8ssandra/cass-operator/tests/util/kubectl"
+
+	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 )
 
 var (
-	testName   = "Test all the things"
-	namespace  = "test-test-all-the-things"
-	dcName     = "dc1"
-	dcYaml     = "../testdata/default-two-rack-two-node-dc.yaml"
-	dcResource = fmt.Sprintf("CassandraDatacenter/%s", dcName)
-	dcLabel    = fmt.Sprintf("cassandra.datastax.com/datacenter=%s", dcName)
-	ns         = ginkgo_util.NewWrapper(testName, namespace)
+	testName       = "Test all the things"
+	namespace      = "test-test-all-the-things"
+	dcName         = "dc1"
+	dcNameOverride = "My_Super_Dc"
+	dcYaml         = "../testdata/default-two-rack-two-node-dc.yaml"
+	dcResource     = fmt.Sprintf("CassandraDatacenter/%s", dcName)
+	dcLabel        = fmt.Sprintf("cassandra.datastax.com/datacenter=%s", api.CleanupForKubernetes(dcNameOverride))
+	ns             = ginkgo_util.NewWrapper(testName, namespace)
 )
 
 func TestLifecycle(t *testing.T) {
@@ -86,10 +89,12 @@ var _ = Describe(testName, func() {
 			ns.WaitForDatacenterOperatorProgress(dcName, "Updating", 60)
 			ns.WaitForDatacenterConditionWithTimeout(dcName, "ScalingUp", string(corev1.ConditionFalse), 1200)
 			// Ensure that when 'ScaleUp' becomes 'false' that our pods are in fact up and running
-			Expect(len(ns.GetDatacenterReadyPodNames(dcName))).To(Equal(4))
+			Expect(len(ns.GetDatacenterReadyPodNames(api.CleanupForKubernetes(dcNameOverride)))).To(Equal(4))
 
 			ns.ExpectDoneReconciling(dcName)
 			ns.WaitForDatacenterReady(dcName)
+
+			ns.ExpectDatacenterNameStatusUpdated(dcName, dcNameOverride)
 
 			step = "stopping the dc"
 			json = "{\"spec\": {\"stopped\": true}}"
@@ -106,7 +111,7 @@ var _ = Describe(testName, func() {
 				FormatOutput(json)
 			ns.WaitForOutputAndLog(step, k, "4", 20)
 
-			ns.WaitForDatacenterToHaveNoPods(dcName)
+			ns.WaitForDatacenterToHaveNoPods(api.CleanupForKubernetes(dcNameOverride))
 
 			step = "resume the dc"
 			json = "{\"spec\": {\"stopped\": false}}"
@@ -125,7 +130,7 @@ var _ = Describe(testName, func() {
 			wg.Add(1)
 			go func() {
 				k = kubectl.Logs("-f").
-					WithLabel("statefulset.kubernetes.io/pod-name=cluster1-dc1-r1-sts-0").
+					WithLabel(fmt.Sprintf("statefulset.kubernetes.io/pod-name=cluster1-%s-r1-sts-0", api.CleanupForKubernetes(dcNameOverride))).
 					WithFlag("container", "cassandra")
 				output, err := ns.Output(k)
 				Expect(err).ToNot(HaveOccurred())