+++ tests/upgrade_operator/upgrade_operator_suite_test.go
@@ -102,7 +102,7 @@ var _ = Describe(testName, func() {
 			// give the operator a minute to reconcile and update the datacenter
 			time.Sleep(1 * time.Minute)
 
-			ns.WaitForDatacenterReadyWithTimeouts(dcName, 1200, 1200)
+			ns.WaitForDatacenterReady(dcName)
 
 			ns.ExpectDoneReconciling(dcName)
 
@@ -126,7 +126,7 @@ var _ = Describe(testName, func() {
 			ns.ExecAndLog(step, k)
 
 			ns.WaitForDatacenterOperatorProgress(dcName, "Updating", 60)
-			ns.WaitForDatacenterReadyWithTimeouts(dcName, 1200, 1200)
+			ns.WaitForDatacenterReady(dcName)
 			ns.WaitForDatacenterReadyPodCount(dcName, 3)
 
 			ns.ExpectDoneReconciling(dcName)